Searched Products appear in pages now
When you search for a category of items they will be displayed 5 at a time. A next and previous button will appear at the top for you to navigate between pages. These buttons disappear when there isn't a next or previous page.

Still need to add some boundary checks and display errors to user.